# openapi.yaml
openapi: 3.0.0
info:
  title: Garudanih API - OpenAPI 3.0
  version: 0.0.1
servers:
  - description: Production
    url: https://api-ticket.up.railway.app/
  - description: Development
    url: http://localhost:8000/
...
paths:
...
  /v1/user/current:
    get:
      security:
      - bearerAuth: []
      tags:
        - Users
      summary: Get current User
      description: Get current user who logged in
      operationId: CurrentUser
      parameters:
        - in: header
          name: token
          description: The token from who's login
          required: true
          schema:
            type: string
            example: eyJhbGciOiJIUzI1NiJ9.eyJSb2xlIjoiQWRtaW4iLCJJc3N1ZXIiOiJJc3N1ZXIiLCJVc2VybmFtZSI6IkphdmFJblVzZSIsImV4cCI6MTY2OTUyOTg4NiwiaWF0IjoxNjY5NTI5ODg2fQ.dS6B2S9bjN5mHA8Zu5iKgRnQKmwRKv8QKl7CzPJGdys
      responses:
        "200":
          $ref: "#/components/responses/DataUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: User don't have permisson because unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "UNAUTHORIZED"
              errors:
                type: string
                example: "You don't have permisson, please login!"
...
    DataUser:
      description: contain all data user
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "SUCCESS"
              data:
                type: object
                properties:
                  users:
                    $ref: "#/components/schemas/Users"
...
  schemas:
    Users:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: "kiki@gmail.com"
        password:
          type: string
          example: "$2b$10$eN9y23azrH5pNbE/gAWqdeYEr3IJ7I7vIAfGUCzDaIpErR2mJhewe"
        role:
          type: string
          example: "member"
        isExist:
          type: boolean
          example: true
        updateAt:
          type: string
          example: "2022-11-27T06:37:22.652Z"
        createdAt:
          type: string
          example: "2022-11-27T06:37:22.652Z"
        name:
          type: string
          example: "kiki pamela"
        image:
          type: string
          example: null
        phone:
          type: string
          example: null
        birth:
          type: string
          example: null
        deletedAt:
          type: string
          example: "2022-11-27T06:37:22.652Z"
...